FIX: use get_target_matches to get today_utc
move get_target_matches to db_manage.py
use get_target_matches to get today_utc in get_user_current_mate
