+++ models.py
@@ -42,7 +42,7 @@ class Match(db.Model):
     __tablename__ = 'matches'
 
     index = db.Column(db.Integer, primary_key=True)
-    match_day = db.Column(db.DateTime(timezone=True), default=datetime.utcnow)
+    match_day = db.Column(db.DateTime(timezone=True), default=datetime.utcnow())
     users = db.relationship(User, secondary=user_identifier, backref='matches')
     activity_index = db.Column(db.Integer, db.ForeignKey('activities.index'))
     activity = db.relationship("Activity")
